fix: cache_download_command

+++ src/constants.rs
@@ -4,7 +4,7 @@ pub const NO: &str = "No";
 /// Default page size: 20
 /// Max page size: 500 (experimentally verified)
 /// Must be explicitly set in order to take effect
-pub const GAMMA_EVENTS_PAGE_SIZE: i32 = 500;
+pub const GAMMA_EVENTS_PAGE_SIZE: usize = 500;
 
 pub const GAMMA_EVENTS_KEYSPACE: &str = "gamma_events";
 pub const CLOB_MARKET_RESPONSES_KEYSPACE: &str = "clob_market_responses";
